server/rag/requirements.txt
langchain
langchain-community
langchain-core
langchain-cohere
langchain-openai
rank_bm25
ragas
langsmith
instructor
python-dotenv
gradio
pydantic
